Move HexToBytesIter to new iter module
In preparation for adding a `BytesToHexIter` type move the
`HexToBytesIter` to its own module. This is a code organization module
so make it private and re-export the type at the crate root as it
currently is.

Refactor only, no logic changes.
