+++ src/lib.rs
@@ -42,10 +42,12 @@ extern crate alloc;
 pub mod buf_encoder;
 pub mod display;
 mod error;
+mod iter;
 pub mod parse;
 
 pub use display::DisplayHex;
-pub use parse::{FromHex, HexToArrayError, HexToBytesError, HexToBytesIter};
+pub use iter::HexToBytesIter;
+pub use parse::{FromHex, HexToArrayError, HexToBytesError};
 
 /// Reexports of extension traits.
 pub mod exts {
